+++ drift_dev/lib/src/analysis/resolver/drift/table.dart
@@ -19,7 +19,7 @@ import 'find_dart_class.dart';
 
 class DriftTableResolver extends LocalElementResolver<DiscoveredDriftTable> {
   static final RegExp _enumRegex =
-      RegExp(r'^enum\((\w+)\)$', caseSensitive: false);
+      RegExp(r'^enum(name)?\((\w+)\)$', caseSensitive: false);
 
   DriftTableResolver(super.file, super.discovered, super.resolver, super.state);
 
@@ -54,10 +54,11 @@ class DriftTableResolver extends LocalElementResolver<DiscoveredDriftTable> {
       AnnotatedDartCode? defaultArgument;
 
       final typeName = column.definition?.typeName;
-      final enumMatch =
+
+      final enumIndexMatch =
           typeName != null ? _enumRegex.firstMatch(typeName) : null;
-      if (enumMatch != null) {
-        final dartTypeName = enumMatch.group(1)!;
+      if (enumIndexMatch != null) {
+        final dartTypeName = enumIndexMatch.group(2)!;
         final imports = file.discovery!.importDependencies.toList();
         final dartClass = await findDartClass(imports, dartTypeName);
 
@@ -72,6 +73,7 @@ class DriftTableResolver extends LocalElementResolver<DiscoveredDriftTable> {
             (msg) => reportError(
                 DriftAnalysisError.inDriftFile(column.definition ?? stmt, msg)),
             dartClass.classElement.thisType,
+            type == DriftSqlType.int ? EnumType.intEnum : EnumType.textEnum,
           );
         }
       }